FEATURE: WPI now configurable via yaml file - weatherPleasantness.yaml
